remove watcher
